refactor(data-model): decompose Fabric hierarchies into by-class dirs
Create two new source directories holding one file per class:

- `src/fabric-instances/` for the `FabricInstance` subclass hierarchy:
  `ExplicitTagValue`, `ProblematicValue`, `UnknownValue` (moved), plus
  `FabricNativeWrapper`, `FabricError`, `FabricMap`, `FabricSet`,
  `FabricRegExp` (split out of `fabric-native-instances.ts`).
- `src/fabric-primitives/` for the `FabricPrimitive` subclass hierarchy:
  `FabricBytes`, `FabricHash` (moved), plus `FabricEpochNsec`,
  `FabricEpochDays` (split out of `fabric-epoch.ts`).

Each directory gets an `index.ts` barrel, exported from `deno.json` as
`./fabric-instances` and `./fabric-primitives` so e.g.
`import { FabricEpochNsec } from "@commonfabric/data-model/fabric-primitives"`
resolves.

The abstract roots `FabricSpecialObject`, `FabricInstance`, and
`FabricPrimitive` stay in `interface.ts` (shared, dependency-free leaf);
the intermediate abstracts `ExplicitTagValue` and `FabricNativeWrapper`
move with their hierarchies.

The three non-class helpers (`isConvertibleNativeInstance`, `UNSAFE_KEYS`,
`errorClassFromType`) stay in a residual `fabric-native-instances.ts`,
still exported as `./fabric-native-instances`. `FabricError` imports
`UNSAFE_KEYS` + `errorClassFromType` back from that residual file.

Class bodies move verbatim; only import paths are repointed. The split
`fabric-epoch` / `fabric-native-instances` per-class `deno.json`
specifiers are dropped in favor of the barrels; cross-package importers
(memory, runner) repoint to the barrel, including the dynamic
`typeof import("...").FabricEpochNsec` form in `runner/builder/types.ts`.
