+++ packages/data-model/deno.json
@@ -28,11 +28,12 @@
     "./BaseReconstructionContext": "./src/BaseReconstructionContext.ts",
     "./deep-freeze": "./src/deep-freeze.ts",
     "./EmptyReconstructionContext": "./src/EmptyReconstructionContext.ts",
-    "./ExplicitTagValue": "./src/ExplicitTagValue.ts",
-    "./FabricBytes": "./src/FabricBytes.ts",
-    "./fabric-epoch": "./src/fabric-epoch.ts",
-    "./FabricHash": "./src/FabricHash.ts",
+    "./ExplicitTagValue": "./src/fabric-instances/ExplicitTagValue.ts",
+    "./FabricBytes": "./src/fabric-primitives/FabricBytes.ts",
+    "./FabricHash": "./src/fabric-primitives/FabricHash.ts",
+    "./fabric-instances": "./src/fabric-instances/index.ts",
     "./fabric-native-instances": "./src/fabric-native-instances.ts",
+    "./fabric-primitives": "./src/fabric-primitives/index.ts",
     "./fabric-type-tags": "./src/fabric-type-tags.ts",
     "./fabric-value": "./src/fabric-value.ts",
     "./fabric-value-legacy": "./src/fabric-value-legacy.ts",
@@ -44,10 +45,10 @@
     "./JsonEncodingContext": "./src/JsonEncodingContext.ts",
     "./json-type-handlers": "./src/json-type-handlers.ts",
     "./native-type-tags": "./src/native-type-tags.ts",
-    "./ProblematicValue": "./src/ProblematicValue.ts",
+    "./ProblematicValue": "./src/fabric-instances/ProblematicValue.ts",
     "./schema-hash": "./src/schema-hash.ts",
     "./schema-utils": "./src/schema-utils.ts",
-    "./UnknownValue": "./src/UnknownValue.ts",
+    "./UnknownValue": "./src/fabric-instances/UnknownValue.ts",
     "./value-debug": "./src/value-debug.ts",
     "./value-hash": "./src/value-hash.ts"
   }
