+++ packages/data-model/src/JsonEncodingContext.ts
@@ -6,22 +6,20 @@ import {
   type ReconstructionContext,
   type SerializationContext,
 } from "./interface.ts";
-import { ExplicitTagValue } from "./ExplicitTagValue.ts";
+import { ExplicitTagValue } from "./fabric-instances/ExplicitTagValue.ts";
 import { deepFreeze } from "./deep-freeze.ts";
-import { UnknownValue } from "./UnknownValue.ts";
-import { ProblematicValue } from "./ProblematicValue.ts";
+import { UnknownValue } from "./fabric-instances/UnknownValue.ts";
+import { ProblematicValue } from "./fabric-instances/ProblematicValue.ts";
 import {
   createDefaultRegistry,
   type JsonWireValue,
   type TypeHandlerCodec,
   type TypeHandlerRegistry,
 } from "./json-type-handlers.ts";
-import {
-  FabricError,
-  FabricMap,
-  FabricRegExp,
-  FabricSet,
-} from "./fabric-native-instances.ts";
+import { FabricError } from "./fabric-instances/FabricError.ts";
+import { FabricMap } from "./fabric-instances/FabricMap.ts";
+import { FabricRegExp } from "./fabric-instances/FabricRegExp.ts";
+import { FabricSet } from "./fabric-instances/FabricSet.ts";
 import { TAGS } from "./fabric-type-tags.ts";
 import { utf8SortedKeysOf } from "@commonfabric/utils/utf8";
