+++ packages/data-model/src/fabric-instances/index.ts
@@ -1,0 +1,8 @@
+export { ExplicitTagValue } from "./ExplicitTagValue.ts";
+export { ProblematicValue } from "./ProblematicValue.ts";
+export { UnknownValue } from "./UnknownValue.ts";
+export { FabricNativeWrapper } from "./FabricNativeWrapper.ts";
+export { FabricError, type FabricErrorState } from "./FabricError.ts";
+export { FabricMap } from "./FabricMap.ts";
+export { FabricSet } from "./FabricSet.ts";
+export { FabricRegExp } from "./FabricRegExp.ts";
