+++ packages/data-model/src/fabric-primitives/index.ts
@@ -1,0 +1,4 @@
+export { FabricBytes } from "./FabricBytes.ts";
+export { FabricHash } from "./FabricHash.ts";
+export { FabricEpochNsec } from "./FabricEpochNsec.ts";
+export { FabricEpochDays } from "./FabricEpochDays.ts";
