+++ packages/data-model/src/json-type-handlers.ts
@@ -4,10 +4,11 @@ import {
   type FabricValue,
   type ReconstructionContext,
 } from "./interface.ts";
-import { ExplicitTagValue } from "./ExplicitTagValue.ts";
-import { ProblematicValue } from "./ProblematicValue.ts";
-import { FabricEpochDays, FabricEpochNsec } from "./fabric-epoch.ts";
-import { FabricBytes } from "./FabricBytes.ts";
+import { ExplicitTagValue } from "./fabric-instances/ExplicitTagValue.ts";
+import { ProblematicValue } from "./fabric-instances/ProblematicValue.ts";
+import { FabricEpochDays } from "./fabric-primitives/FabricEpochDays.ts";
+import { FabricEpochNsec } from "./fabric-primitives/FabricEpochNsec.ts";
+import { FabricBytes } from "./fabric-primitives/FabricBytes.ts";
 import { TAGS } from "./fabric-type-tags.ts";
 import {
   fromBase64url,
