+++ packages/data-model/src/schema-hash.ts
@@ -4,7 +4,7 @@
  */
 
 import type { JSONSchema, JSONSchemaObj } from "@commonfabric/api";
-import { FabricHash } from "./FabricHash.ts";
+import { FabricHash } from "./fabric-primitives/FabricHash.ts";
 import { SchemaAndHash } from "./SchemaAndHash.ts";
 import { toDeepFrozenSchema } from "./schema-utils.ts";
 import { hashOf, hashStringOf } from "./value-hash.ts";
