+++ packages/data-model/test/json-encoding.test.ts
@@ -6,7 +6,7 @@ import {
   seemsLikeJsonEncodedFabricValue,
   valueFromJson,
 } from "../src/json-encoding.ts";
-import { FabricError } from "../src/fabric-native-instances.ts";
+import { FabricError } from "../src/fabric-instances/FabricError.ts";
 import type { FabricValue } from "../src/fabric-value.ts";
 import { BaseReconstructionContext } from "../src/BaseReconstructionContext.ts";
