+++ packages/runner/src/builder/types.ts
@@ -358,9 +358,9 @@ export interface BuilderFunctionsAndConstants {
   FabricPrimitive:
     typeof import("@commonfabric/data-model/fabric-value").FabricPrimitive;
   FabricEpochNsec:
-    typeof import("@commonfabric/data-model/fabric-epoch").FabricEpochNsec;
+    typeof import("@commonfabric/data-model/fabric-primitives").FabricEpochNsec;
   FabricEpochDays:
-    typeof import("@commonfabric/data-model/fabric-epoch").FabricEpochDays;
+    typeof import("@commonfabric/data-model/fabric-primitives").FabricEpochDays;
   FabricHash: typeof import("@commonfabric/data-model/FabricHash").FabricHash;
 
   // Debug stringifiers
