+++ packages/runner/test/memory-v2-stacked-commit.test.ts
@@ -12,7 +12,7 @@ import {
   jsonFromValue,
   valueFromJson,
 } from "@commonfabric/data-model/json-encoding";
-import { FabricEpochNsec } from "@commonfabric/data-model/fabric-epoch";
+import { FabricEpochNsec } from "@commonfabric/data-model/fabric-primitives";
 import { Identity } from "@commonfabric/identity";
 import type { FabricValue, MIME, URI } from "@commonfabric/memory/interface";
 import {
